add companion tool diff-pdf-page-statistics.py
The tool is meant to help find regressions.

The python script takes the output of mso-test.exe
and compares the imported and exported PDFs
against the authoritative PDF,
showing the correct position in red.

CSV reports are generated to report
the change in red pixels.

Another CSV lists any suspicious anomalies detected.

And finally, comparison images are created,
either on demand or automatically when
the number of red pixels has increased
since a previous milestone.