diff-pdf-page-statistics.py: only force-save PNGs of more-red pages
Previously I saved the ENTIRE comparison if ANY page had more red.
There was SOME value in doing that,
but just browsing through these is the best way to find regressions,
so limit this output only to times when we have more red than before
to avoid wasting a lot of QA visual processing time.

I considered just saving the FIRST page that had more red
(since often a change will cascade to multiple pages)
but it is easy enough to manually skip past the rest of the pages
if you have already identified that this document is problematic.
Otherwise I might be hiding a major regression behind a false positive.