diff-pdf-page-statistics.py: overlay red with blue prev and grayscale MS_ORIG
Previously, we only looked at the different between now and previously.
But nothing indicated what it SHOULD look like.

I first tried blue under red,
but that wasn't nice because you can't easily distinguish
between a regression and an existing problem.

With blue over top of red, the color of blue indicates an existing problem,
(or a good fix when no red - in the case when overlays are always saved)
and the color red indicates either a regression or a partial move
either towards or away from the ideal represented by the gray.

I think this change will be quite helpful.