Improved the shell scripts.
The generate-bmp-diffs.sh is now more versatile. The user has to enter the directory the pdf lives.
They are also able to set where the output off the diffs will be.
They also can set some options like max pages and enable minor differences.
This means that the user no longer has to use pdf-to-bmp.sh has the current script handles that, with the given arguments.

Signed-off-by: Elliot Freebairn <elliot.freebairn@collabora.com>