Replacing python imagemagick code with the c++ program
Changing the arguments passed to c++ so that it reflects what was passed previously in the python script.

Added the ability to compare round tripped MSO pdfs.

Fixing some referencing and memory bugs.

Starting to work on moving the logging from shell to csv files, like justin does in the script.

Still alot more work to do, but nearly there.

Signed-off-by: Elliot Freebairn <elliot.freebairn@collabora.com>