Improving sensitivity to subtle pixel changes
A new colour is added, dark yellow.

It indicates that a pixel is located on an edge and is significantly different

The dark yellow pixel may be a false positive.

However it has a much higher chance of being significantly different.

Also reduced the radius of pixels blurred near edges/text

 This reduces the amount of pixels ignored for differences.

Signed-off-by: Elliot Freebairn <elliot.freebairn@collabora.com>