Feat: a new csv is written to containg pdf anomalies
Initially, the program will convert all pages to images, so we can record the number of pages for each pdf.

If there a mismtach in number of pages, that information is written to a separate csv file

After that, the pages are trimmed so that there are only max_pages, or the minimum number of pages from one of the PDF'S in the set.

Signed-off-by: Elliot Freebairn <elliot.freebairn@collabora.com>